addon/edit: introduce indentlist addon
This addon provides two functions for smartly indenting or unindenting
list items.

If the cursor is located right after the list bullet, pressing Tab
indents the list bullet by one tab; pressing Shift-Tab unindents the
list bullet by one tab.

Pressing Tab or Shift-Tab while the cursor is located within the text
that follows a list bullet yields the regular behavior.

This addon is used in the next commit in order to be able to (un)indent
lists in Markdown (issue #4355).

Signed-off-by: Joel Porquet <joel@porquet.org>